registar/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
mysqlclient==2.2.4
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
python-dotenv==1.0.1
pyzbar==0.1.9
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
